# Make geojson optional, use geo-types for geometry representation.