--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,8 +13,13 @@ keywords = ["contour", "polygon", "isoring", "marching-squares", "geojson"]
 license = "MIT/Apache-2.0"
 
 [dependencies]
-geojson = ">=0.16, <=0.24"
+geojson = { version = ">=0.16, <=0.24", optional = true }
+geo-types= { version = "0.7.0" }
 lazy_static = "1.0"
 serde_json = "^1.0"
 rustc-hash = "1.0"
 slab = "0.4"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]