@mthh 2022-09-23
In fact I may be AFK the next few days, so in order not to let this PR linger I will merge it now, fix the small typo and change the "value" field to "threshold" in the geojson representation in subsequent commits, update changelog and readme and make a new release today.

Thanks again for the very useful contribution !